% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PltsPre.R
\name{PltsPre}
\alias{PltsPre}
\title{PltsPre is a function to predict platelet cells  in single-cell RNA sequencing (scRNA-seq) data.}
\usage{
PltsPre(suerat_object, cutoff = 0.6911)
}
\arguments{
\item{suerat_object}{A Seurat object that metadata contains nFeature_RNA.}

\item{cutoff}{Numeric. Cutoff value for binary prediction. Default is 0.6911.}
}
\value{
A Seurat object with predicted cell labels and scores.
}
\description{
PltsPre is a function to predict platelet cells  in single-cell RNA sequencing (scRNA-seq) data.
}
\examples{
# Example usage of the function
data("pbmc_small")
pbmc_small <- PltsPre(pbmc_small)

}
